[ADVAPI32][USERENV][SERVICES] Unify source for vista registry functions
- In advapi32_vista compile the code from advapi32/reg/reg.c
- Export RegCopyTreeW from advapi32_vista
- Remove duplicated code / hacks from advap32_vista, userenv and services
