+++ base/system/services/CMakeLists.txt
@@ -23,6 +23,6 @@ add_executable(services ${SOURCE} services.rc)
 target_link_libraries(services ${PSEH_LIB})
 
 set_module_type(services win32gui UNICODE)
-add_importlibs(services userenv user32 advapi32 rpcrt4 msvcrt kernel32 ntdll)
+add_importlibs(services userenv user32 advapi32 advapi32_vista rpcrt4 msvcrt kernel32 ntdll)
 add_pch(services services.h SOURCE)
 add_cd_file(TARGET services DESTINATION reactos/system32 FOR all)
