+++ dll/win32/advapi32_vista/advapi32_vista.spec
@@ -1,4 +1,5 @@
 
+@ stdcall RegCopyTreeW(ptr wstr ptr)
 @ stdcall RegDeleteTreeA(long str)
 @ stdcall RegDeleteTreeW(long wstr)
 @ stdcall RegSetKeyValueW(long wstr wstr long ptr long)
